requirements.txt
google-api-python-client==2.132.0
djangorestframework==3.13
django==3.2
oauth2client==4.1.3
django-celery-beat
django-celery-results
redis
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
